dspy-gepa-deidentification/pyproject.toml
[project]
name = "dspy-gepa-deidentification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy-ai>=3.0.3",
    "gepa>=0.0.7",
    "ipykernel>=7.0.1",
]
